+++ R/demographic_model.R
@@ -202,11 +202,27 @@ dm.addParameter <- function(dm, par.name, lower.boundary, upper.boundary, fixed.
 #'            possibly containing one or more previously created parameter
 #'            names.
 #' @param group     For genomic features, different groups can be created.
-#' @return          The extended demographic model.
+#' @param variance Set to a value different from 0 to introduce variation in the
+#'                 the parameter value for different loci. The 
+#'                 variation follows a discrete gamma distribution with mean equal to
+#'                 the value provided as \code{parameter}, and variance as given
+#'                 here. Can also be set to a previously 
+#'                 created parameter, or an expression based on parameters. 
+#' @param var.classes This sets the number of classes used in the discrete gamma
+#'                 distribution described in \code{variance}. Fewer classes should
+#'                 increase performance, but leads to harder approximation of the
+#'                 gamma distribution.
+#' @param zero.inflation If used, a zero inflated discrete gamma distribution is 
+#'                 used to model the variablility between loci. This parameter
+#'                 should evluate to the percent of loci for which the parameter
+#'                 is 0. The values for all other loci will be drawn from the
+#'                 discretized gamma distribution.
+#' @return         The extended demographic model.
 addFeature <- function(dm, type, parameter=NA,
                        lower.range=NA, upper.range=NA, fixed.value=NA, par.new=T,
                        pop.source=NA, pop.sink=NA,
-                       time.point=NA, group=0) {
+                       time.point=NA, group=0, 
+                       variance=0, var.classes=5, zero.inflation = 0) {
   
   if (missing(par.new))     par.new <- T
   if (missing(parameter))   parameter <- NA
@@ -224,9 +240,31 @@ addFeature <- function(dm, type, parameter=NA,
   checkType(pop.sink,    c("num",  "s"), T, T)
   checkType(time.point,  c("char", "s"), T, T)
   checkType(group,       c("num",  "s"), T, T)
+  checkType(variance,    c("s"), T, T)
+  checkType(var.classes, c("num",  "s"), T, T)
 
-  if (par.new) dm <- dm.addParameter(dm, parameter, lower.range, upper.range, fixed.value)
+  if (par.new) dm <- dm.addParameter(dm, parameter, lower.range, 
+                                     upper.range, fixed.value)
 
+  if (variance != 0) {
+    if (dm.getSubgroupNumber(dm) > 1) 
+      stop("Variation is only supported for one parameter per model")
+    dm <- dm.addSubgroups(dm, var.classes, group = group, 
+                          zero_inflation = zero.inflation)
+    if (zero.inflation == 0) {
+      parameter <- paste0('calcDiscreteGammaRate(subgroup, ', parameter, ', ',
+                          variance, ', ', var.classes, ')')
+    } else {
+      parameter <- paste0('ifelse(subgroup == 1, 0, ', 
+                          'calcDiscreteGammaRate(subgroup-1, ', parameter, ', ',
+                          variance, ', ', var.classes, '))')
+    }
+      
+  } else if (zero.inflation != 0) {
+    dm <- dm.addSubgroups(dm, 1, group = group, zero_inflation = zero.inflation)
+    parameter <- paste0('ifelse(subgroup == 1, 0, ', parameter, ")")
+  }
+  
   # Append the feature
   dm <- appendToFeatures(dm = dm,
                          type = type,
@@ -503,8 +541,7 @@ dm.setLociLength <- function(dm, loci.length, group=0) {
 #' @return The number of loci in the group
 #' @export
 dm.getLociNumber <- function(dm, group=1) {
-  ln <- searchFeature(dm, type='loci.number', group=group)
-  as.integer(ln$parameter)
+  as.integer(searchFeature(dm, type='loci.number', group=group)$parameter)
 }
 
 #' Gets how long the loci in a group are
@@ -554,6 +591,20 @@ dm.getLociLength <- function(dm, group=1) {
 #'            created. You can also use R expression here, so "2*theta"
 #'            or "5*theta+2*tau" (if tau is another parameter) will also
 #'            work (also it does not make much sense).
+#' @param variance Set to a value different from 0 to introduce variation in the
+#'                 the parameter value for different loci. The 
+#'                 variation follows a discrete gamma distribution with mean equal to
+#'                 the value provided as \code{parameter}, and variance as given
+#'                 here. Can also be set to a previously 
+#'                 created parameter, or an expression based on parameters. 
+#' @param var.classes This sets the number of classes used in the discrete gamma
+#'                 distribution described in \code{variance}. Fewer classes should
+#'                 increase performance, but leads to harder approximation of the
+#'                 gamma distribution.
+#' @param group    Group of loci for with this feature is added. 0 means that
+#'                 the feature applies to all groups, and 1 is the default group.
+#'                 Set to 1 or an greater integer to set this feature only for 
+#'                 the corresponding group of loci. 
 #' @return    The demographic model with mutation.
 #' @export
 #'
@@ -573,7 +624,8 @@ dm.getLociLength <- function(dm, group=1) {
 #' dm <- dm.addParameter(dm, "theta", 0.01, 5)
 #' dm <- dm.addMutation(dm, par.new=FALSE, parameter="2*log(theta)+1")
 dm.addMutation <- function(dm, lower.range, upper.range, fixed.value,
-                           par.new=T, new.par.name="theta", parameter) {
+                           par.new=T, new.par.name="theta", parameter,
+                           group = 0, variance = 0, var.classes = 5) {
 
   if ( missing(lower.range) & missing(upper.range) & 
        missing(fixed.value) & missing(parameter) ) {
@@ -583,7 +635,9 @@ dm.addMutation <- function(dm, lower.range, upper.range, fixed.value,
 
   if (par.new) parameter <- new.par.name
   dm <- addFeature(dm, "mutation", parameter, lower.range, upper.range,
-                   fixed.value, par.new=par.new, time.point=NA)
+                   fixed.value, par.new=par.new, time.point=NA, 
+                   variance = variance, var.classes = var.classes,
+                   group = group)
   return(dm)
 }
 
@@ -664,6 +718,20 @@ dm.getSampleSize <- function(dm, group.nr=NULL) {
 #'            created. You can also use R expression here, so "2*rho"
 #'            or "5*rho+2*tau" (if tau is another parameter) will also
 #'            work (also it does not make much sense).
+#' @param variance Set to a value different from 0 to introduce variation in the
+#'                 the parameter value for different loci. The 
+#'                 variation follows a discrete gamma distribution with mean equal to
+#'                 the value provided as \code{parameter}, and variance as given
+#'                 here. Can also be set to a previously 
+#'                 created parameter, or an expression based on parameters. 
+#' @param var.classes This sets the number of classes used in the discrete gamma
+#'                 distribution described in \code{variance}. Fewer classes should
+#'                 increase performance, but leads to harder approximation of the
+#'                 gamma distribution.
+#' @param group    Group of loci for with this feature is added. 0 means that
+#'                 the feature applies to all groups, and 1 is the default group.
+#'                 Set to 1 or an greater integer to set this feature only for 
+#'                 the corresponding group of loci. 
 #' @return    The demographic model with recombination
 #' @export
 #'
@@ -673,11 +741,14 @@ dm.getSampleSize <- function(dm, group.nr=NULL) {
 #' dm <- dm.addRecombination(dm, fixed=20)
 #' dm <- dm.addMutation(dm, 1, 20)
 dm.addRecombination <- function(dm, lower.range, upper.range, fixed.value,
-                                par.new=T, new.par.name="rho", parameter) {
+                                par.new=T, new.par.name="rho", parameter, 
+                                group = 0, variance = 0, var.classes = 5) {
 
   if (par.new) parameter <- new.par.name
   dm <- addFeature(dm, "recombination", parameter, lower.range, upper.range,
-                   fixed.value, par.new=par.new, time.point="0")
+                   fixed.value, par.new=par.new, time.point="0",
+                   group = group, 
+                   variance = variance, var.classes = var.classes)
   return(dm)
 }
 
@@ -1107,8 +1178,8 @@ dm.setMutationModel <- function(dm, mutation.model,
 #-------------------------------------------------------------------
 # dm.addMutationRateHeterogenity
 #-------------------------------------------------------------------
-#' Allows the mutation rate on different sites to vary according to 
-#' a Gamma Distribution.
+#' Allows the mutation rate on different sites within one locus to 
+#' vary according to a Gamma Distribution.
 #'
 #' This function adds a Gamma distributed rate heterogeneity as implemented 
 #' in 'seq-gen' to the model.
@@ -1260,6 +1331,7 @@ dm.addOutgroup <- function(dm, separation.time) {
 # This function is highly experimental. Don't use it yet.
 dm.addPositiveSelection <- function(dm, min.strength, max.strength, fixed.strength, 
                          par.new=T, new.par.name="s", parameter, 
+                         variance = 0, var.classes = 5, fraction.neutral = 0,
                          population, at.time, group=0) {
 
   checkType(population, c("num",  "s"), T, F)
@@ -1268,7 +1340,9 @@ dm.addPositiveSelection <- function(dm, min.strength, max.strength, fixed.streng
   if (par.new) parameter <- new.par.name
 
   dm <- addFeature(dm, "pos.selection", parameter, min.strength, max.strength,
-                   fixed.strength, par.new, population, NA, at.time, group)
+                   fixed.strength, par.new, population, NA, at.time, group,
+                   variance = variance, var.classes = var.classes, 
+                   zero.inflation = fraction.neutral)
 
   return(dm)
 }
@@ -1431,4 +1505,45 @@ scaleDemographicModel <- function(dm, scaling.factor) {
                      group)
   }
   return(dm)
+}
+
+dm.addSubgroups <- function(dm, subgroup_number, zero_inflation = 0, group = 0) {
+  checkType(subgroup_number, c('num', 's'))
+  checkType(group, c('num', 's'))
+  checkType(zero_inflation, c('s'), FALSE)
+  
+  if (zero_inflation != 0) {
+    dm <- addFeature(dm, 'zero_inflation', as.character(zero_inflation),
+                     group = group, par.new = FALSE)
+  }
+  
+  addFeature(dm, 'subgroups', as.character(subgroup_number), 
+             group = group, par.new = FALSE)
+}
+
+dm.getSubgroupNumber <- function(dm, group = 1, with_zero_inflasion = TRUE) {
+  number <- searchFeature(dm, 'subgroups', group = group)$parameter
+  
+  if (length(number) == 0) number <- 1
+  else number <- as.numeric(number)
+  
+  if(with_zero_inflasion) {
+    if (!is.na(dm.getZeroInflation(dm, group = group))) number <- number + 1
+  }
+  
+  number
+}
+
+dm.getZeroInflation <- function(dm, group = 1) {
+  zi <- searchFeature(dm, 'zero_inflation', group = group)
+  if (nrow(zi) == 0) return(NA)
+  zi$parameter
+}
+
+# Calulates the rate for the different gamma subgroups.
+# Uses the median of the group, because the mean is difficult to approximate.
+calcDiscreteGammaRate <- function(subgroup, mean, var, n_subgroups) {
+  stopifnot(var >= 0)
+  if (var == 0) return(rep(mean, length(subgroup)))
+  qgamma((2*subgroup-1)/(2*n_subgroups), mean^2/var, mean/var)
 }